Add pedestrian lights
Pedestrian lights are added with the use of an extra timer and an
interrupt routine assigned to it, so its control can take place
asynchronously. To make up for the 8bit timer and to meet the need
of a relatively long period (~250ms) for writing out to the leds,
an additional counter was used on top of the timer interrupt. That
way the led write is triggered only after a certain count of timer
overflows.